create curry function
create curry function in lib.js which takes a function and a number as
an argument and returns a function. That function takes max. two
arguments and pass them as an input to the function which is passed as
an argument to curry function along with the number argument and the
return of that function is taken as the output.